Add initial support for version handling (#76)
This makes it easier to work with data where either you already know the
version associated with it, or if you have a function that can
dynamically get the version. The first time I wrote code like this was
in PyOBO, but I've also written similar code many times in different
places in combination with pystow, so it makes sense to enable it
upstream here.

This PR just adds support for the join and ensure functions, but in a
second step can be added to all functions that transitively call them.